win-system build errors

+++ win-system/Environment.cpp
@@ -113,7 +113,7 @@ bool Environment::getSpecialFolderPath(int specialFolderId, StringStorage *out)
   TCHAR path[MAX_PATH + 1];
 
 
-  
+
 
   if (SHGetSpecialFolderPath(NULL, &path[0], csidl, TRUE) == TRUE) {
     out->setString(&path[0]);
@@ -170,7 +170,7 @@ bool Environment::getCurrentModuleFolderPath(StringStorage *out)
   size_t lastPos = out->findLast(_T('\\'));
 
   if (lastPos != (size_t)-1) {
-    out->getSubstring(out, 0, max(lastPos - 1, 0));
+    out->getSubstring(out, 0, max(static_cast<int>(lastPos -1), 0));
   }
 
   return true;
@@ -207,7 +207,7 @@ void Environment::restoreWallpaper(LogWriter *log)
   }
 
   int ret = 1;
-   
+
 SystemParametersInfo(SPI_SETMENUANIMATION, 1,&ret, 0);
 SystemParametersInfo(SPI_SETSELECTIONFADE, 1, &ret, 0);
 SystemParametersInfo(SPI_SETTOOLTIPANIMATION, 1, &ret, 0);
@@ -216,35 +216,35 @@ SystemParametersInfo(SPI_SETDROPSHADOW, 1, &ret, 0);
 SystemParametersInfo(SPI_SETDRAGFULLWINDOWS, 1, &ret, 0);
 
 
-  
+
 // enable animation
 
 ANIMATIONINFO animinf;
 //UINT prevAnim;
 
-animinf.cbSize = sizeof(ANIMATIONINFO); 
-SystemParametersInfo(SPI_GETANIMATION,sizeof(ANIMATIONINFO),&animinf,0); 
-animinf.iMinAnimate = 1; 
-SystemParametersInfo(SPI_SETANIMATION,sizeof(ANIMATIONINFO),&animinf,0); 
+animinf.cbSize = sizeof(ANIMATIONINFO);
+SystemParametersInfo(SPI_GETANIMATION,sizeof(ANIMATIONINFO),&animinf,0);
+animinf.iMinAnimate = 1;
+SystemParametersInfo(SPI_SETANIMATION,sizeof(ANIMATIONINFO),&animinf,0);
 
 }
 
 void Environment::disableWallpaper(LogWriter *log)
 {
 
-  
+
   log->info(_T("Try to disable wallpaper"));
   Impersonator imp(log);
   AutoImpersonator ai(&imp, log);
-  
+
   Sleep(300); // dirty hack, wait for mirage driver
 
   if (SystemParametersInfo(SPI_SETDESKWALLPAPER, 0, _T(""), 0) == 0) {
     throw SystemException(_T("Cannot disable desktop wallpaper"));
   }
-  
+
   // disable visual effects
-  
+
   if (SystemParametersInfo(SPI_SETMENUANIMATION, 0, NULL, 0) == 0) {
     throw SystemException(_T("Cannot disable menu animation"));
   }
@@ -264,22 +264,22 @@ void Environment::disableWallpaper(LogWriter *log)
     throw SystemException(_T("Cannot disable drag fullwindow"));
   }
 
-  
+
 // disable animation
 
 
 ANIMATIONINFO animinf;
 UINT prevAnim;
 
-animinf.cbSize = sizeof(ANIMATIONINFO); 
-SystemParametersInfo(SPI_GETANIMATION,sizeof(ANIMATIONINFO),&animinf,0); 
-prevAnim = animinf.iMinAnimate; 
+animinf.cbSize = sizeof(ANIMATIONINFO);
+SystemParametersInfo(SPI_GETANIMATION,sizeof(ANIMATIONINFO),&animinf,0);
+prevAnim = animinf.iMinAnimate;
 
-if(prevAnim) 
-{ 
-animinf.iMinAnimate = 0; 
-SystemParametersInfo(SPI_SETANIMATION,sizeof(ANIMATIONINFO),&animinf,0); 
-} 
+if(prevAnim)
+{
+animinf.iMinAnimate = 0;
+SystemParametersInfo(SPI_SETANIMATION,sizeof(ANIMATIONINFO),&animinf,0);
+}
 
 //////
 
@@ -341,7 +341,7 @@ void Environment::simulateCtrlAltDel(LogWriter *log)
   }else{
 	simulateCtrlAltDelUnderVista(log);
   }
-  
+
 }
 
 void Environment::startControlPanel(LogWriter *log)
@@ -409,38 +409,38 @@ bool Environment::isAeroOn(LogWriter *log)
 
 bool Environment::initReboot()
 {
-    HANDLE hToken; 
-   TOKEN_PRIVILEGES tkp; 
- 
-   // Get a token for this process. 
- 
-   if (!OpenProcessToken(GetCurrentProcess(), 
-        TOKEN_ADJUST_PRIVILEGES | TOKEN_QUERY, &hToken)) 
-      return( FALSE ); 
- 
-   // Get the LUID for the shutdown privilege. 
- 
-   LookupPrivilegeValue(NULL, SE_SHUTDOWN_NAME, 
-        &tkp.Privileges[0].Luid); 
- 
-   tkp.PrivilegeCount = 1;  // one privilege to set    
-   tkp.Privileges[0].Attributes = SE_PRIVILEGE_ENABLED; 
- 
-   // Get the shutdown privilege for this process. 
- 
-   AdjustTokenPrivileges(hToken, FALSE, &tkp, 0, 
-        (PTOKEN_PRIVILEGES)NULL, 0); 
- 
-   if (GetLastError() != ERROR_SUCCESS) 
-      return FALSE; 
- 
-   // Shut down the system and force all applications to close. 
- 
-   if (!ExitWindowsEx(EWX_REBOOT | EWX_FORCE, 
+    HANDLE hToken;
+   TOKEN_PRIVILEGES tkp;
+
+   // Get a token for this process.
+
+   if (!OpenProcessToken(GetCurrentProcess(),
+        TOKEN_ADJUST_PRIVILEGES | TOKEN_QUERY, &hToken))
+      return( FALSE );
+
+   // Get the LUID for the shutdown privilege.
+
+   LookupPrivilegeValue(NULL, SE_SHUTDOWN_NAME,
+        &tkp.Privileges[0].Luid);
+
+   tkp.PrivilegeCount = 1;  // one privilege to set
+   tkp.Privileges[0].Attributes = SE_PRIVILEGE_ENABLED;
+
+   // Get the shutdown privilege for this process.
+
+   AdjustTokenPrivileges(hToken, FALSE, &tkp, 0,
+        (PTOKEN_PRIVILEGES)NULL, 0);
+
+   if (GetLastError() != ERROR_SUCCESS)
+      return FALSE;
+
+   // Shut down the system and force all applications to close.
+
+   if (!ExitWindowsEx(EWX_REBOOT | EWX_FORCE,
                SHTDN_REASON_MAJOR_OPERATINGSYSTEM |
                SHTDN_REASON_MINOR_UPGRADE |
-               SHTDN_REASON_FLAG_PLANNED)) 
-      return FALSE; 
+               SHTDN_REASON_FLAG_PLANNED))
+      return FALSE;
 
    //shutdown was successful
    return TRUE;
@@ -455,7 +455,7 @@ StringStorage folder,path;
 if (Environment::getCurrentModuleFolderPath(&path)) {
     StringStorage entry(_T("Software\\Microsoft\\")
                         _T("Windows\\CurrentVersion\\RunOnce\\"));
-	folder.format(_T("%s\\konturvnc.exe"),path.getString()); 
+	folder.format(_T("%s\\konturvnc.exe"),path.getString());
 	folder.quoteSelf();
 	RegistryKey regKey(HKEY_CURRENT_USER, entry.getString());
 	regKey.setValueAsString(_T("konturVNC"), folder.getString());
